requirements.txt
openai>=1.12.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
